Consolidate utils sprawl
